Check if building solution has failed and if so, return exit code 1

+++ scripts/linux/buildBranch.sh
@@ -32,26 +32,41 @@ if [ ! -d "${BUILD_BINARIESDIRECTORY}" ]; then
 fi
 
 echo Building all solutions in repo
-find $ROOTFOLDER -type f -name $SLN_SUFFIX | while read soln; do
+RES=0
+
+while read soln; do
     echo Building Solution - $soln
     $DOTNET_ROOT_PATH/dotnet clean --output $BUILD_BINARIESDIRECTORY $soln
     $DOTNET_ROOT_PATH/dotnet restore $soln
     $DOTNET_ROOT_PATH/dotnet build --output $BUILD_BINARIESDIRECTORY $soln
-done
+    if [ $? -gt 0 ]; then
+        RES=1
+    fi
+done < <(find $ROOTFOLDER -type f -name $SLN_SUFFIX)
+
+if [ $RES -ne 0 ]; then
+    exit $RES
+fi
 
 echo Publishing all required solutions in repo
 rm -fr $PUBLISH_FOLDER
 
-find $ROOTFOLDER -type f -name $CSPROJ_SUFFIX -exec grep -l "<OutputType>Exe</OutputType>" {} + | while read proj; do
+while read proj; do
     echo Publishing Solution - $proj
     filename=$(basename "$proj")
     proj_publish_dir="${filename%.*}"
     $DOTNET_ROOT_PATH/dotnet publish -f netcoreapp2.0 -o $PUBLISH_FOLDER/$proj_publish_dir $proj
+    if [ $? -gt 0 ]; then
+        RES=1
+    fi
     # This is a workaround for dotnet publish issue when it adds netstandard1.1 to dependencies path
     echo Remove netstandard1.1 from deps.json
     sed -i 's/lib\/netstandard1.1\///g' $PUBLISH_FOLDER/$proj_publish_dir/$proj_publish_dir.deps.json
-done
+
+done < <(find $ROOTFOLDER -type f -name $CSPROJ_SUFFIX -exec grep -l "<OutputType>Exe</OutputType>" {} +)
 
 echo Copying $SRC_DOCKER_DIR to $PUBLISH_FOLDER/docker
 rm -fr $PUBLISH_FOLDER/docker
 cp -r $SRC_DOCKER_DIR $PUBLISH_FOLDER
+
+exit $RES
